LibGL: Introduce Texture base class for all texture types
