+++ Userland/Libraries/LibGL/SoftwareRasterizer.h
@@ -9,7 +9,7 @@
 #include "DepthBuffer.h"
 #include "GL/gl.h"
 #include "GLStruct.h"
-#include "Tex/Texture.h"
+#include "Tex/Texture2D.h"
 #include <AK/OwnPtr.h>
 #include <LibGfx/Bitmap.h>
 #include <LibGfx/Vector4.h>
@@ -31,7 +31,7 @@ class SoftwareRasterizer final {
 public:
     SoftwareRasterizer(const Gfx::IntSize& min_size);
 
-    void submit_triangle(const GLTriangle& triangle, const Texture& texture);
+    void submit_triangle(const GLTriangle& triangle, const Texture2D& texture);
     void resize(const Gfx::IntSize& min_size);
     void clear_color(const FloatVector4&);
     void clear_depth(float);
